# Improved CPU performance by almost 13ms (37.48ms -> 24.74ms) on the RosCon level