--- a/Gems/Atom/RPI/Code/Source/RPI.Public/Pass/Pass.cpp
+++ b/Gems/Atom/RPI/Code/Source/RPI.Public/Pass/Pass.cpp
@@ -1302,7 +1302,6 @@ namespace AZ
 
         void Pass::FrameBegin(FramePrepareParams params)
         {
-            AZ_PROFILE_SCOPE(RPI, "Pass::FrameBegin() - %s", m_path.GetCStr());
             AZ_RPI_BREAK_ON_TARGET_PASS;
 
             bool earlyOut = !IsEnabled();
@@ -1335,10 +1334,7 @@ namespace AZ
 
             // FrameBeginInternal needs to be the last function be called in FrameBegin because its implementation expects 
             // all the attachments are imported to database (for example, ImageAttachmentPreview)
-            {
-                AZ_PROFILE_SCOPE(RPI, "Pass::FrameBeginInternal()");
-                FrameBeginInternal(params);
-            }
+            FrameBeginInternal(params);
             
             // readback attachment with output state
             UpdateReadbackAttachment(params, false);