--- a/Gems/Atom/RPI/Code/Source/RPI.Public/ViewportContext.cpp
+++ b/Gems/Atom/RPI/Code/Source/RPI.Public/ViewportContext.cpp
@@ -141,8 +141,16 @@ namespace AZ
 
         void ViewportContext::OnBeginPrepareRender()
         {
+            AZ_PROFILE_FUNCTION(RPI);
             ViewportContextNotificationBus::Event(GetName(), &ViewportContextNotificationBus::Events::OnRenderTick);
             ViewportContextIdNotificationBus::Event(GetId(), &ViewportContextIdNotificationBus::Events::OnRenderTick);
+        }
+
+        void ViewportContext::OnEndPrepareRender()
+        {
+            AZ_PROFILE_FUNCTION(RPI);
+            ViewportContextNotificationBus::Event(GetName(), &ViewportContextNotificationBus::Events::WaitForRender);
+            ViewportContextIdNotificationBus::Event(GetId(), &ViewportContextIdNotificationBus::Events::WaitForRender);
         }
 
         AZ::Name ViewportContext::GetName() const