--- a/Gems/Atom/Tools/AtomToolsFramework/Code/Source/PreviewRenderer/PreviewRenderer.cpp
+++ b/Gems/Atom/Tools/AtomToolsFramework/Code/Source/PreviewRenderer/PreviewRenderer.cpp
@@ -80,7 +80,7 @@ namespace AtomToolsFramework
         m_renderPipeline->SetDefaultView(m_view);
 
         // Don't render the pipeline until a capture request
-        m_renderPipeline->RemoveFromRenderTick();
+        m_renderPipeline->AddToRenderTickOnce();
 
         m_state.reset(new PreviewRendererIdleState(this));