--- a/Gems/AtomLyIntegration/ImguiAtom/Code/Source/ImguiAtomSystemComponent.cpp
+++ b/Gems/AtomLyIntegration/ImguiAtom/Code/Source/ImguiAtomSystemComponent.cpp
@@ -95,6 +95,13 @@ namespace AZ
 #endif
         }
 
+        void ImguiAtomSystemComponent::WaitForRender()
+        {
+#if defined(IMGUI_ENABLED)
+            ImGui::ImGuiManagerBus::Broadcast(&ImGui::IImGuiManager::WaitForRenderToFinish);
+#endif
+        }
+
         void ImguiAtomSystemComponent::OnViewportSizeChanged([[maybe_unused]] AzFramework::WindowSize size)
         {
 #if defined(IMGUI_ENABLED)