Improved CPU performance with the following optimizations:
- Removed a redundant call to UpdateVisibility at the end of the frame (~3ms)
- Switched culling to use jobs instead of task graph (~2.5ms)
- Switch the tools render pipeline to be off when created since it gets manually turned on and then off when it needs to do a capture (~0.5ms)
- Switch ImGui update to be multithreaded (up to 5ms when CPU profiler is running)

Those were the big ones if I'm not forgetting any. Other changes here are small things like removing perf markers that are too small to show up in profiling, adding perf markers for tasks that are definitely big enough to profile,